requirements.txt
uvicorn
django
djangorestframework
pydantic 
alembic 
psycopg2
itsdangerous
pydantic[email]
web3
